[trusty_app_manifest] Added new START_PORT config option
This option allows applications to specify ports that when connected
to will start the application if it was not already running. The option
takes a port name and port flags that determine the types of connections
allowed. Additionally, a new MGMT_FLAG is available for applications that
do not wish to be started at boot (and will presumably be started later
with a port connection)

Change-Id: Ica1e61a4482674d1b21ae74d675fa6f1c082fa66